Move the shared tile grid-layout helpers out of Directory
